Ported HttpParserBench to BenchmarkDotNet
